# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/sources": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get a list of sources",
                "description": "Fetches a list of all sources",
                "operationId": "14187994c89fe56b7c1d044c4558766b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of sources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Source"
                                            }
                                        },
                                        "page_context": {
                                            "properties": {
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "string",
                                                    "example": "20"
                                                },
                                                "has_more_page": {
                                                    "type": "string",
                                                    "example": "false"
                                                },
                                                "sort_column": {
                                                    "type": "string",
                                                    "example": "created_at"
                                                },
                                                "sort_order": {
                                                    "type": "string",
                                                    "example": "ASC"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sources"
                ],
                "summary": "Create a new source",
                "description": "Stores a new source in the database",
                "operationId": "38bc7ddd15345a798bf9d216610b91ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "title",
                                    "rss_feed_link"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example Source"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Example Title"
                                    },
                                    "rss_feed_link": {
                                        "type": "string",
                                        "example": "http://example.com/rss"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Source created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Source"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sources/{id}": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get source details",
                "description": "Fetches a single source by its ID",
                "operationId": "9595ed26f1cff3f6c4d85b0e5e9ad5a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the source",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Source"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Source": {
                "title": "Source",
                "description": "A source entity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Example Source"
                    },
                    "rss_feed_link": {
                        "type": "string",
                        "example": "http://example.com/rss"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-07T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-07T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}